Change of LICENSE to GNU LGPLv3
Change of LICENSE to GNU LGPLv3. This change is granted by the old license (QaPL v0.2 sec. [4]) without the permission of the contributors.

The reason for the change are some requests about using code from ClearURLs in other project, e.g. an iOS App and an application to route URLs to configured browsers. The old license is not so friendly to use ClearURLs in other projects. Also I think ClearURLs should be a real free software.

All new versions are licensed under the GNU LGPLv3. For all old versions: As the Original Copyright Holder I hereby grant a dual-licensed, under QaPL v0.2 and GNU LGPLv3, for all version of ClearURLs older or equal to version 1.5a.